Document value($X) in message
This doc update is really crappy but I was having massive writer's block
because even basic things that would have made it better (like using full
rules instead of "formula") were really unappealing to me. I wonder if
the reason is that I wasn't confident in how to structure this or if this
was even the right place to document this feature, given that we don't seem
to explain how to use metavariables in messages. My only hope is that this
is understandable to the tech writers who can help me make it better <3.