Fix sample-ci-configs.md's uploading findings file (#758)
Fixed the "Alternate job that uploads findings to GitHub Advanced Security Dashboard" CI files to actually upload the SARIF files to the dashboard. With the current command they do not upload any findings as a SARIF file is never created. This is to be more inline with 

https://semgrep.dev/docs/semgrep-ci/running-semgrep-ci-with-semgrep-app/#setting-up-security-dashboards-for-github-and-gitlab 

which does upload the findings.