Add references to enabling comments docs
I found my own KB in a search today (lol) and realized it was weird from that perspective that it doesn't link to the docs for enabling these. 

Logically speaking for this KB to be useful, you need to already have them enabled, but you might still land on it in other circumstances. 😄