We should not tell people to use --output
The --output is a bit redundant with shell redirection
so we should not advise it. It was added a long time ago
and we're not even sure anymore if it's that useful.
People in theory can use it to post the result to an URL
but even that I'm not sure it's used. In any case
when a user just want to save the output to a file, he
can simply use shell redirection.

test plan:
wait for green CI checks