Update description and troubleshooting for longer scans
Folks sometimes see this message if a scan is long as well so this clarifies that the scan results will still be accepted if the scan is running.

I removed the last troubleshooting bullet and related statement since they are incorrect. Modifying --timeout only changes how much time Semgrep spends on a rule/file combo. There is no global timeout for Semgrep, only whatever the running system enforces.