Update description and troubleshooting for longer scans (#1845)
* Update description and troubleshooting for longer scans

Folks sometimes see this message if a scan is long as well so this clarifies that the scan results will still be accepted if the scan is running.

I removed the last troubleshooting bullet and related statement since they are incorrect. Modifying --timeout only changes how much time Semgrep spends on a rule/file combo. There is no global timeout for Semgrep, only whatever the running system enforces.

* Apply suggestions from TW review

Co-authored-by: Katie Horne <katie.horne@semgrep.com>

* Match usage from heading change

Co-authored-by: Finn Ellis <finn@semgrep.com>

* Commit half of Finn's suggestion :)

---------

Co-authored-by: Katie Horne <katie.horne@semgrep.com>
Co-authored-by: Finn Ellis <finn@semgrep.com>