Replace X placeholder with environment variable
The contributing docs that explain how to add and update a programming
language use a placeholder `X` to refer to the name of the language.
This commit updates the docs to use `$PL` as the placeholder instead.

Using `$PL` makes it easier to copy and run the code samples in the
documentation, and `$PL` is more descriptive than simply `$X`.