Explicitly use `semgrep` switch in `opam env` commands
The `opam switch create` command in the instructions for contributing
creates an opam switch named `semgrep`, but later commands don't
explicitly refer to a switch when setting up the environment in order
to use it, which may cause unintended behaviour if more than one switch
exists. Explicitly refer to the `semgrep` switch in invocations of
`opam env`.