fix(perf-principles): don't reference specific log file name (#2389)
The perf principles document references enabling `--verbose` and checking a file named `ci.log`.  By default, verbose and debug output is sent to standard out, not a file with that name.  This patch simply changes where the user should expect to actually see the output.

As an additional nit: `--debug` will also output the relevant information, so I clarified the flags that will do so.

### Please ensure

- [x] A subject matter expert (SME) reviews the content
- [x] A technical writer reviews the content or PR
- [x] This change has no security implications or else you have pinged the security team
- [x] Redirects are added if the PR changes page URLs
- [x] If you have changed any header tag links (doc/#this-kind-of-anchor), update all instances of that link